requirements.txt
python-dotenv
Pillow
google-generativeai
google-ai-generativelanguage